+++ CHANGELOG.md
@@ -2,9 +2,21 @@
 
 ## 1.32.5
 
+### Improvements
+
+- Add a PATCH annotation endpoint ([#1904](../../pull/1904))
+- Add an option to merge dicom files in the same series and study and folder ([#1912](../../pull/1912))
+
+### Changes
+
+- Tweak a few source priorities ([#1907](../../pull/1907), [#19098](../../pull/19098))
+
 ### Bug Fixes
 
 - Always use api root where appropriate ([#1902](../../pull/1902))
+- Fix listing updated time in customized annotation lists ([#1903](../../pull/1903))
+- Guard isGeopatial when called on an image GDAL can open but not read ([#1906](../../pull/1906))
+- Improve handling of large annotation elements ([#1908](../../pull/1908))
 
 ## 1.32.4